x86 GetSlotLifetimes: handle pushed pointer arg transitions
Add tracking for GcTransitionPointer PUSH/POP/KILL actions and
StackDepthTransition in x86 GetSlotLifetimes. Pushed GC pointer
arguments now emit SP-relative stack slot lifetimes from their
push offset to their pop/kill offset.

Also handle register PUSH/POP actions for stack depth tracking
(non-pointer register pushes affect the depth but don't create
GC slot lifetimes).

+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Contracts/Contracts/GCInfo/GCInfoDecoder.cs
@@ -1216,10 +1216,10 @@ internal class GcInfoDecoder<TTraits> : IGCInfoDecoder where TTraits : IGCInfoTr
         // TODO(stackref): The native FindSafePoint uses binary search (NarrowSafePointSearch)
         // when numSafePoints > 32. This is a performance optimization only — no correctness impact.
         // Linear scan through safe point offsets from the saved position
-        for (uint i = 0; i < _safePoints.Count; i++)
+        for (int i = 0; i < _safePoints.Count; i++)
         {
             if (_safePoints[i] == codeOffset)
-                return i;
+                return (uint)i;
             if (_safePoints[i] > codeOffset)
                 break;
         }
